+++ lighthouse-core/test/lib/traces/metrics-evts-test.js
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2016 Google Inc. All rights reserved.
+ * Copyright 2017 Google Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,28 +27,29 @@ const dbwResults = require('../../fixtures/dbw_tester-perf-results.json');
 
 /* eslint-env mocha */
 describe('metrics events class', () => {
-  let evts;
   it('exposes metric definitions', () => {
     assert.equal(Metrics.metricsDefinitions.length, 8, 'eight metrics not exposed');
   });
 
   it('generates fake trace events', () => {
-    evts = new Metrics(dbwTrace.traceEvents, dbwResults.audits).generateFakeEvents();
+    const evts = new Metrics(dbwTrace.traceEvents, dbwResults.audits).generateFakeEvents();
 
     const metricsWithoutNavstart = Metrics.metricsDefinitions.length - 1;
+    // The trace events must come in pairs, thus the `2 * n`
     assert.equal(evts.length, 2 * metricsWithoutNavstart, 'All expected fake events not created');
 
     const definitionsWithoutEvents = Metrics.metricsDefinitions
         .filter(metric => metric.id !== 'navstart')
         .filter(metric => !evts.find(e => e.name === metric.name));
-    assert.deepEqual(definitionsWithoutEvents, [], 'metrics are missing fake events');
+    assert.strictEqual(definitionsWithoutEvents.length, 0, 'metrics are missing fake events');
 
     const eventsWithoutDefinitions = evts
         .filter(evt => !Metrics.metricsDefinitions.find(metric => metric.name === evt.name));
-    assert.deepEqual(eventsWithoutDefinitions, [], 'fake events without a metric definition');
+    assert.strictEqual(eventsWithoutDefinitions.length, 0, 'fake events w/o a metric definition');
   });
 
   it('generates fake trace events that are valid', () => {
+    const evts = new Metrics(dbwTrace.traceEvents, dbwResults.audits).generateFakeEvents();
     const vizCompleteEvts = evts.filter(e => e.name.includes('Visually Complete 100'));
     assert.equal(vizCompleteEvts.length, 2, 'Two visually complete 100% events not found');
     assert.equal(vizCompleteEvts[0].id, vizCompleteEvts[1].id, 'UT trace ids don\'t match');
@@ -62,24 +63,6 @@ describe('metrics events class', () => {
       assert.equal(typeof e.name, 'string');
       assert.equal(typeof e.ph, 'string');
       assert.equal(typeof e.args, 'object');
-    });
-  });
-});
-
-describe('assetSaver prepareAssets', () => {
-  it('adds fake events to trace', () => {
-    const countEvents = trace => trace.traceEvents.length;
-    const mockArtifacts = {
-      traces: {
-        defaultPass: dbwTrace
-      },
-      requestScreenshots: () => Promise.resolve([]),
-    };
-    const beforeCount = countEvents(dbwTrace);
-    return assetSaver.prepareAssets(mockArtifacts, dbwResults).then(preparedAssets => {
-      const afterCount = countEvents(preparedAssets[0].traceData);
-      const metricsSansNavStart = Metrics.metricsDefinitions.length - 1;
-      assert.equal(afterCount, beforeCount + (2 * metricsSansNavStart), 'unexpected event count');
     });
   });
 });
